Access middleware support for alternate logging
Updated the access middleware module to enable alternate loggers
besides passing a single line string to a logging function to
support just one format and one type of output.  With this
system, the existing Logger in access.Logger is still available,
but additional loggers that support services like e.g. datadog,
loggly, or Amazon Cloudwatch would now be possible to create.

Created a new CustomLogger function that generates middleware for
logging given a function that accepts the request and response
writer objects and the elapsed time for processing the request.
The function can then extract whatever data it wants about the
processing of the request from what is available on the request
and response writer and the elapsed time provided and output it
anywhere it likes.

CustomLogger was created by separating the middleware functionality
and elapsed time calculation from the actual logging, which is now
handled by the function passed in.

The original Logger was then updated to use the CustomLogger by
removing the middleware handling and elapsed time calculation parts
of it and taking the output generation and putting that into a
function of type LogWriterFunc.  Logger then just calls
CustomLogger with the new output generation function as an argument
and returns the routing.Handler CustomLogger generates.